renamed: viewalltestidisplay file
